feat(phase1): implement Postgres legacy-migration path (round 3 finding #1)
Closes the last round-3 P2: the Phase 1 plan (Task 10) documents Postgres as a
supported legacy-import target, but PostgresStore had none of the bulk-import or
parity-read methods, so migrate import-sqlite blocked PG. Implemented the full
path and validated it against a real PostgreSQL via testcontainers.

PostgresStore now implements the complete BulkImportPort + ParityReadPort:
- begin/commit/rollback_import_transaction hold one pooled connection across the
  whole import so every bulk_insert_* shares the transaction.
- disable/enable_triggers toggle the auto-seed trigger (ALTER TABLE ... TRIGGER
  USER); reset_sequences advances the jobs/job_status_history SERIAL sequences
  past the explicitly imported ids.
- 10 bulk_insert_* methods via executemany, plus read_all_rows/count_rows.

Two real bugs the live PG run surfaced (mypy could not):
- Legacy timestamps are TEXT but PG columns are TIMESTAMPTZ; $N::timestamptz
  makes asyncpg infer the param as timestamptz and reject the string. Cast as
  $N::text::timestamptz so PG parses the legacy ISO text.
- PG schema used REAL (float4) for ml_gate_score and the cost columns, losing
  precision vs SQLite's double (0.6691 -> 0.6690999865531921) and failing the
  parity checksum. Changed those five columns to DOUBLE PRECISION.

Also: parity canonicalizer now normalizes native datetime values (PG returns
aware datetimes; SQLite returns text) so checksums match across backends.
migrate.py accepts SQLiteStore | PostgresStore (guard removed).

Tests: tests/e2e/test_legacy_import_postgres.py (import + parity, id/sequence
preservation). Full PG suite: 78 passed against real PG. SQLite: 247 passed,
lint+mypy clean.

+++ migrations/versions/0001_initial_schema.py
@@ -53,7 +53,7 @@ def upgrade() -> None:
         tech_required TEXT,
         role_type TEXT,
         yoe_min INTEGER,
-        ml_gate_score REAL,
+        ml_gate_score DOUBLE PRECISION,
         ml_gate_result TEXT,
         ml_gate_fail_reason TEXT,
         ml_gate_at TIMESTAMPTZ,
@@ -79,7 +79,7 @@ def upgrade() -> None:
         ),
         stage_a_error TEXT,
         stage_a_model TEXT,
-        stage_a_cost_usd REAL,
+        stage_a_cost_usd DOUBLE PRECISION,
         stage_a_prompt_hash TEXT,
         stage_a_resume_hash TEXT,
         stage_b_verdict TEXT CHECK (
@@ -97,7 +97,7 @@ def upgrade() -> None:
         ),
         stage_b_error TEXT,
         stage_b_model TEXT,
-        stage_b_cost_usd REAL,
+        stage_b_cost_usd DOUBLE PRECISION,
         stage_b_prompt_hash TEXT,
         stage_b_resume_hash TEXT,
         created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
@@ -124,7 +124,7 @@ def upgrade() -> None:
         stage_a_scored INTEGER DEFAULT 0,
         stage_b_scored INTEGER DEFAULT 0,
         jobs_scored INTEGER DEFAULT 0,
-        total_llm_cost_usd REAL DEFAULT 0.0,
+        total_llm_cost_usd DOUBLE PRECISION DEFAULT 0.0,
         errors INTEGER DEFAULT 0,
         finished_at TIMESTAMPTZ
     )
@@ -214,7 +214,7 @@ def upgrade() -> None:
     op.execute("""
     CREATE TABLE cost_ledger (
         day TEXT PRIMARY KEY,
-        spent_usd REAL NOT NULL DEFAULT 0.0,
+        spent_usd DOUBLE PRECISION NOT NULL DEFAULT 0.0,
         calls INTEGER NOT NULL DEFAULT 0,
         last_updated TIMESTAMPTZ NOT NULL DEFAULT now()
     )
